Nicer comments, naming changes. Group posts integration tests into single group.
